# app/docs/swagger.yaml
definitions:
...
  domain.AuthHome:
    properties:
      sevenDayUser:
        $ref: '#/definitions/domain.UserData'
        description: 7日信息
      todayInsertUser:
        $ref: '#/definitions/domain.UserData'
        description: 当日用户信息
      totalUser:
        description: 总用户
        type: integer
    type: object
...
  domain.FbEnv:
    properties:
      createTime:
        type: string
      devEnv:
        description: 开发环境
        type: string
      envType:
        description: '环境类型: 1-环境变量 2-系统变量'
        type: integer
      id:
        type: integer
      isDel:
        type: integer
      key:
        description: 环境变量key
        type: string
      proEnv:
        description: 生产环境
        type: string
      updateTime:
        type: string
    type: object
...
  domain.Home:
    properties:
      homeApi:
        $ref: '#/definitions/domain.HomeApi'
      homeAuth:
        $ref: '#/definitions/domain.HomeAuth'
      homeDataSource:
        $ref: '#/definitions/domain.HomeDataSource'
      homeOss:
        $ref: '#/definitions/domain.HomeOss'
    type: object
  domain.HomeApi:
    properties:
      liveQueryTotal:
        description: 实时查询个数
        type: integer
      mutationsTotal:
        description: 更改个数
        type: integer
      queryTotal:
        description: 查询个数
        type: integer
      subscriptionsTotal:
        description: 订阅个数
        type: integer
    type: object
  domain.HomeAuth:
    properties:
      authTotal:
        description: 累计身份验证商
        type: integer
      todayInsertUser:
        description: 今日新增用户
        type: integer
      totalUser:
        description: 累计用户
        type: integer
    type: object
  domain.HomeBulletin:
    properties:
      bulletinType:
        type: integer
      date:
        type: string
      title:
        type: string
    type: object
  domain.HomeDataSource:
    properties:
      CustomerTotal:
        description: 自定义个数
        type: integer
      GraphqlTotal:
        description: graphql个数
        type: integer
      RestTotal:
        description: rest个数
        type: integer
      dbTotal:
        description: db个数
        type: integer
    type: object
  domain.HomeOss:
    properties:
      ossTotal:
        description: oss存储个数
        type: integer
      totalMemory:
        description: 总内存
        type: string
      useMemory:
        description: 已使用内存
        type: string
    type: object
...
  domain.UserData:
    properties:
      count:
        description: 新增数量
        type: integer
      upCount:
        description: 对比前个时间段新增数量
        type: integer
    type: object
...
host: localhost:9123
info:
  contact: {}
  description: FireBoom-server
  title: FireBoom-server Swagger API
  version: "1.0"
paths:
...
  /api/v1/auth/home:
    get:
      consumes:
      - application/json
      description: 查询首页数据
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.AuthHome'
            type: array
        "400":
          description: 查询失败
      tags:
      - auth/home
...
  /api/v1/draft:
    put:
      consumes:
      - application/json
      description: 更新草稿内容
      parameters:
      - description: 内容
        in: formData
        name: content
        required: true
        type: string
      - description: 内容
        in: formData
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 更新成功
        "400":
          description: 更新失败
      tags:
      - draft
  /api/v1/draft/:id:
    delete:
      consumes:
      - application/json
      description: 删除草稿
      parameters:
      - description: 数据源id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - draft
    get:
      consumes:
      - application/json
      description: 根据数据源id获取草稿
      parameters:
      - description: 数据源id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 查询成功
          schema:
            type: string
        "400":
          description: 查询失败
      tags:
      - draft
  /api/v1/env:
    get:
      consumes:
      - application/json
      description: 查询环境变量信息
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.FbEnv'
            type: array
        "400":
          description: 查询失败
      tags:
      - env
    post:
      consumes:
      - application/json
      description: 添加环境变量
      parameters:
      - description: key
        in: formData
        name: key
        required: true
        type: string
      - description: 开发环境
        in: formData
        name: devEnv
        required: true
        type: string
      - description: 生产环境
        in: formData
        name: proEnv
        required: true
        type: string
      - description: 环境类型 1-环境变量 2-系统变量
        in: formData
        name: envType
        required: true
        type: integer
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - env
    put:
      consumes:
      - application/json
      description: 修改环境变量
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      - description: key
        in: formData
        name: key
        required: true
        type: string
      - description: 开发环境
        in: formData
        name: devEnv
        required: true
        type: string
      - description: 生产环境
        in: formData
        name: proEnv
        required: true
        type: string
      - description: 环境类型 1-环境变量 2-系统变量
        in: formData
        name: envType
        required: true
        type: integer
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - env
  /api/v1/env/:id:
    delete:
      consumes:
      - application/json
      description: 删除环境变量
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - env
  /api/v1/env/keys:
    get:
      consumes:
      - application/json
      description: 查询环境变量keys
      parameters:
      - description: key
        in: formData
        name: key
        required: true
        type: string
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.FbEnv'
            type: array
        "400":
          description: 查询失败
      tags:
      - env
  /api/v1/file:
    delete:
      consumes:
      - application/json
      description: 删除文件
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - file
  /api/v1/file/upload:
    post:
      consumes:
      - application/json
      description: 添加保存文件
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - file
...
  /api/v1/home:
    get:
      consumes:
      - application/json
      description: 获取首页数据
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/domain.Home'
        "400":
          description: 查询失败
      tags:
      - home
  /api/v1/home/bulletin:
    get:
      consumes:
      - application/json
      description: 获取公告
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.HomeBulletin'
            type: array
        "400":
          description: 查询失败
      tags:
      - home
...
produces:
- application/json
swagger: "2.0"